Kokoro: install GLFW dependencies
Dawn depends on GLFW, which in turn depends on a number of system
dependencies to be installed. These are not part of the docker image, so
need to be installed on the fly.

Also disable the Kokoro doxygen step as this is still being worked on.

Bug: dawn:1339
Bug: dawn:1358
Bug: tint:1498
Change-Id: I1a33044d11c243fc3809c2b5b4d25f0bfc943d6e
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/86071
Auto-Submit: Ben Clayton <bclayton@google.com>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>