Updates to dawn_version_generator.
This CL fixes the incorrect `path` variable used when the HEAD lookup
fails, the proper path is now emitted.

The call to `get_gitResolvedHead` is wrapped in a try/except and an
exception is treated as if we are not in a git directory. This allows
the build to continue if `refs/head/main` doesn't exist in the local
check.

Change-Id: I49ae66cc5eac99d43387b58d28d356282b8f6783
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/87100
Reviewed-by: Ben Clayton <bclayton@google.com>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Commit-Queue: Dan Sinclair <dsinclair@chromium.org>
Auto-Submit: Dan Sinclair <dsinclair@chromium.org>