Make formatting an error on presubmit
Currently the presubmit will emit a warning if the formatting fails
but this is easy to miss as the bot goes green. This CL changes
the presubmit to make formatting an error so the issue will be
caught and fixedup before landing.

Change-Id: Ifb83d2a7e524fd086e86b47f7518fea1e65f6d3e
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/109563
Kokoro: Kokoro <noreply+kokoro@google.com>
Reviewed-by: Austin Eng <enga@chromium.org>
Commit-Queue: Dan Sinclair <dsinclair@chromium.org>